// AngularClient/AppSrc/app/shared/Controllers.generated.ts
export class Dish implements IDish {
    name?: string | undefined;
    price: number;
    id: string;
    description?: string | undefined;
    imageUrl?: string | undefined;

    constructor(data?: IDish) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.name = data["name"];
            this.price = data["price"];
            this.id = data["id"];
            this.description = data["description"];
            this.imageUrl = data["imageUrl"];
        }
    }

    static fromJS(data: any): Dish {
        let result = new Dish();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["price"] = this.price;
        data["id"] = this.id;
        data["description"] = this.description;
        data["imageUrl"] = this.imageUrl;
        return data; 
    }
}

export interface IDish {
    name?: string | undefined;
    price: number;
    id: string;
    description?: string | undefined;
    imageUrl?: string | undefined;
}